GCP_practice/requirements.txt
google-cloud-storage
google-cloud-firestore
flask
gunicorn
line-bot-sdk
google-cloud-logging